# [mono] Convert Environment.CurrentManagedThreadId to an icall